fix: resolve critical build error and StyleCop warnings (part 1)
**Critical Build Error Fixed:**
- Removed IsToolProfile binding from GameProfileGeneralSettingsView.axaml
- Property doesn't exist on GameProfileSettingsViewModel

**StyleCop Fixes:**
- Added trailing comma to PlayerType enum (SA1413)
- Split ReplayConverters.cs into 3 separate files (SA1402, SA1649):
  - PlayerTypeToColorConverter.cs
  - PlayerTypeToIconConverter.cs
  - ColorNameToBrushConverter.cs
- Added XML documentation to all converter methods (CS1591, SA1600)
- Added trailing commas to all switch expressions in converters (SA1413)

Remaining: Fix ReplayParserService, ReplaySaveService, ReplayMonitoringService warnings
