+++ GenHub/GenHub.Core/Models/Tools/ReplayManager/PlayerType.cs
@@ -18,5 +18,5 @@ public enum PlayerType
     /// <summary>
     /// Observer/spectator.
     /// </summary>
-    Observer
+    Observer,
 }
